+++ src/async_quickcheck.mli
@@ -1,0 +1,6 @@
+open Core_kernel.Std
+open Async_quickcheck_intf
+
+include Quickcheck_async
+
+module Configure (Config : Quickcheck_config) : Quickcheck_async