+++ src/config.mli
@@ -26,9 +26,14 @@ module Max_num_open_file_descrs : sig
 end
 
 module Dump_core_on_job_delay : sig
+  module How_to_dump : sig
+    type t = Default | Call_abort | Call_gcore
+    with sexp
+  end
+
   type watch =
     { dump_if_delayed_by : Time_ns.Span.t
-    ; how_to_dump        : [ `Default | `Call_abort | `Call_gcore ]
+    ; how_to_dump        : How_to_dump.t
     }
   with sexp