+++ src/throttle_unit_tests.ml
@@ -40,7 +40,7 @@ TEST =
 TEST_UNIT =
   let t = create ~continue_on_error:false ~max_concurrent_jobs:1 in
   let i = ref 0 in
-  let _ : unit Deferred.t = enqueue t (fun () -> incr i; Deferred.unit) in
+  let (_ : unit Deferred.t) = enqueue t (fun () -> incr i; Deferred.unit) in
   assert (!i = 0);
   stabilize ();
   assert (!i = 1)